OPTIM: Reduce verbosity and align nuget restore
Use the Nuget restore method recommended in this Microsoft article rather than a call to nuget executable that may not be aware of all the issues associated with restoring packages and dependencies that msbuild is aware of. https://learn.microsoft.com/en-us/nuget/reference/msbuild-targets#restoring-and-building-with-one-msbuild-command

Added the parameter -verbosity:minimal to reduce the long log output that is of little value.