dotnet: write-error on build/test failure
Exit code that is returned currently wasn't being caught by the calling
scripts, as they expect PowerShell scripts to handle native exit codes
themselves. So we do just that - throw an error if any of the tests
failed. The caller's `$ErrorActionPreference` would then be able to
catch it.

Ideally `build-project-*.ps1` scripts should use `$PSNativeCommandUseErrorActionPreference`
instead, but that'd be best done as part of overall refactor to make all
scrips that call native commands use it.