documentation: allow using other branches
The workflow still has to run from main to make simultaneous runs harder
to happen accidentally. The intended way to use this is to run the
workflow multiple times from main, serially, and pass the branch as a
parameter if needed. This is to avoid attempting to push to gh-pages from
multiple jobs simultaneously.